+++ CHANGELOG.md
@@ -99,8 +99,8 @@ Developers may need to change their code
   * Added maskable flags that can be combined when debugging:
     * `DEFAULT` = 1: Basic debug message, espcially initialization
     * `MSG` = 2: Input and output packets, read datastore
-    * `DETAIL` = 4: Details: message dump in hex, xpath parse trees, etc
-    * `EXTRA` = 8: Extra detailed logs
+    * `DETAIL` = 4: xpath parse trees, etc
+    * `EXTRA` = 8: Extra detailed logs, message dump in hex
 * Added `ISO/IEC 10646` encodings to XML parser: `&#[0-9]+;` and `&#[0-9a-fA-F]+;`
 * Added `CLIXON_CLIENT_SSH` to client API to communicate remotely via SSH netconf sub-system